requirements.txt
black==24.10.0
Django==5.1.5
djangorestframework==3.15.2
python-dotenv==1.0.1